fixed typo in tables.sql

+++ Assignment5/sql_scripts/tables.sql
@@ -10,6 +10,6 @@ CREATE TABLE IF NOT EXISTS resources (
 CREATE TABLE IF NOT EXISTS sandwiches (
 id int NOT NULL AUTO_INCREMENT,
 sandwich_size varchar(50) NOT NULL,
-price decimal (5,2) NOT NULL,
+price float NOT NULL,
 PRIMARY KEY (id)
 );
